[BE : DEV] : S07P12B309-210 : add deleteUser API

See merge request s07-webmobile1-sub2/S07P12B309!71

+++ backend/src/main/java/com/ssafy/DTO/projectInfoDTO.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 @Data
 public class projectInfoDTO {
-    int pid;
+    Long pid;
     userSimpleInfoDTO projectOwner;
     List<userSimpleInfoDTO> member;
     LocalDateTime pjtStartDate;
